+++ lunes_cms/cmsv2/admins/__init__.py
@@ -1,3 +1,4 @@
+from .area_admin import AreaAdmin
 from .feedback_admin import FeedbackAdmin
 from .job_admin import JobAdmin
 from .unit_admin import UnitAdmin, UnitWordRelationAdmin
@@ -5,6 +6,7 @@ from .user_admin import LunesUserAdmin
 from .word_admin import WordAdmin
 
 __all__ = [
+    "AreaAdmin",
     "JobAdmin",
     "WordAdmin",
     "UnitAdmin",