+++ lunes_cms/core/settings.py
@@ -571,6 +571,7 @@ REST_FRAMEWORK = {
     "DEFAULT_API_URL": "http://localhost:8080/api/",
     "EXCEPTION_HANDLER": "lunes_cms.api.exception_handler.custom_exception_handler",
     "DEFAULT_THROTTLE_RATES": {
+        "area_registration": "60/hour",
         "installation": "100/min",
         "gdpr": "10/hour",
         "user": "1000/day",
@@ -631,6 +632,7 @@ JAZZMIN_SETTINGS = {
         "cms.GroupAPIKey": "fas fa-key",
         "cms.Feedback": "fas fa-comment",
         "cms.Sponsor": "fas fa-star",
+        "cmsv2.Area": "fa-solid fa-object-group",
         "cmsv2.Job": "fas fa-briefcase",
         "cmsv2.Unit": "fas fa-book",
         "cmsv2.Word": "fab fa-amilia",